- Jorrit fixed a bug in csSector and meshgen. It was casting 	  csMeshGenerator to iMeshGenerator without even knowing the 	  right inheritance structure. This caused meshgen's to not 	  destruct properly (i.e. crash). 	- Jorrit fixed compilation of radixsort on 64-bit.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@23544 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csutil/radixsort.h
@@ -61,7 +61,7 @@ public:
   /**
    * Get the last generated ranks array.
    */
-  inline size_t* GetRanks() const
+  inline uint32* GetRanks() const
   {
     return ranks1;
   }
@@ -139,8 +139,8 @@ private:
   bool DoPass(size_t pass, T* data, size_t size, uint32* histogram);
 
   size_t currentSize;
-  size_t* ranks1;
-  size_t* ranks2;
+  uint32* ranks1;
+  uint32* ranks2;
 
   bool ranksValid;
 };
